+++ packages/crypto/package.json
@@ -51,6 +51,10 @@
       "types": "./src/index.d.ts",
       "import": "./dist/src/index.js"
     },
+    "./ciphers": {
+      "types": "./dist/src/ciphers/index.d.ts",
+      "import": "./dist/src/ciphers/index.js"
+    },
     "./hmac": {
       "types": "./dist/src/hmac/index.d.ts",
       "import": "./dist/src/hmac/index.js"
@@ -58,6 +62,10 @@
     "./keys": {
       "types": "./dist/src/keys/index.d.ts",
       "import": "./dist/src/keys/index.js"
+    },
+    "./webcrypto": {
+      "types": "./dist/src/webcrypto/index.d.ts",
+      "import": "./dist/src/webcrypto/index.js"
     }
   },
   "eslintConfig": {
@@ -106,7 +114,7 @@
     "./dist/src/keys/ed25519/index.js": "./dist/src/keys/ed25519/index.browser.js",
     "./dist/src/keys/rsa/index.js": "./dist/src/keys/rsa/index.browser.js",
     "./dist/src/keys/secp256k1/index.js": "./dist/src/keys/secp256k1/index.browser.js",
-    "./dist/src/webcrypto.js": "./dist/src/webcrypto.browser.js"
+    "./dist/src/webcrypto/webcrypto.js": "./dist/src/webcrypto/webcrypto.browser.js"
   },
   "sideEffects": false
 }
