+++ packages/crypto/src/keys/ecdh/index.browser.ts
@@ -3,7 +3,7 @@ import { concat as uint8ArrayConcat } from 'uint8arrays/concat'
 import { equals as uint8ArrayEquals } from 'uint8arrays/equals'
 import { toString as uint8ArrayToString } from 'uint8arrays/to-string'
 import { base64urlToBuffer } from '../../util.js'
-import webcrypto from '../../webcrypto.js'
+import webcrypto from '../../webcrypto/index.js'
 import type { Curve } from './index.js'
 import type { ECDHKey, ECDHKeyPair, JWKEncodedPrivateKey, JWKEncodedPublicKey } from '../interface.js'
