+++ packages/crypto/src/webcrypto/index.ts
@@ -1,0 +1,3 @@
+import webcrypto from './webcrypto.js'
+
+export default webcrypto
